# Reduction kernel optimization