**Description**
Optimize reduction kernel code by moving loads from global memory before computation.
Add CMake option to build CUDA code with --generate-line-info option.

Results from running float reduce_square_sum() 1000 times:

number of elements | baseline average ms | updated average ms | baseline/updated
-- | -- | -- | --
1024 | 0.005959 | 0.005421 | 1.09931
2048 | 0.008346 | 0.007856 | 1.062425
4096 | 0.008491 | 0.007929 | 1.070858
8192 | 0.008609 | 0.008151 | 1.056164
1048576 | 0.012841 | 0.011321 | 1.134248
1073741824 | 8.312 | 5.12766 | 1.621012

**Motivation and Context**
Address performance regression observed in training throughput tests from #5639.