@weixingzhang 2020-12-09
Have the training throughput tests been merged in ORT repo?