@edgchen1 2020-12-09
> Have the training throughput tests been merged in ORT repo? If not, I am wondering whether these tests can be merged in ORT repo to catch the regression for the future change.

Yes, these are the tests that show up on the performance dashboard.
https://github.com/microsoft/onnxruntime/blob/master/tools/ci_build/github/azure-pipelines/orttraining-linux-gpu-perf-test-ci-pipeline.yml